unit_1/requirements.txt
kafka-python==2.0.2
avro-python3==1.10.2